+++ packages/calendar/src/Calendar/__tests__/Calendar.test.tsx
@@ -188,4 +188,34 @@ describe('Calendar', () => {
     fireEvent.press(calendarPressTitleYear); // range years
     expect(toJSON()).toMatchSnapshot();
   });
+  it('renders correctly dottedDates', () => {
+    const {toJSON} = render(
+      <Calendar
+        isClear={true}
+        initialRange={{fromDate: '2022-01-16', toDate: '2022-01-19'}}
+        isShowToday={false}
+        dottedDates={[
+          new Date('2022-02-08'),
+          new Date('2022-02-28'),
+          new Date('2022-02-16'),
+        ]}
+      />,
+    );
+    expect(toJSON()).toMatchSnapshot();
+  });
+  it('renders correctly dottedDates cc', () => {
+    const {toJSON} = render(
+      <Calendar
+        isClear={true}
+        initialRange={{fromDate: '2022-01-16', toDate: '2022-01-19'}}
+        isShowToday={false}
+        dottedDates={[
+          new Date('2022-02-08'),
+          new Date('2022-02-28'),
+          new Date('2022-02-16'),
+        ]}
+      />,
+    );
+    expect(toJSON()).toMatchSnapshot();
+  });
 });