+++ packages/calendar/src/Calendar/functions.ts
@@ -16,6 +16,7 @@ export const getDateForCalendar = (date: Date): string => {
   const d = `${date.getDate() < 10 ? 0 : ''}${date.getDate()}`;
   return `${yr}-${month}-${d}`;
 };
+
 const getStyleToday = (colorToday: colorElem): MarkingProps => {
   return {
     startingDay: true,
@@ -114,6 +115,19 @@ export const getAllDatesBetween = (
   };
 
   return {dates: datesForCalendar, fromDate, toDate, lengthDateRange};
+};
+
+export const getDottedDates = (dots: Date[], dotColor: string) => {
+  const datesForCalendar: IMarkedTypes = {};
+
+  for (const dot of dots) {
+    datesForCalendar[getDateForCalendar(dot)] = {
+      marked: true,
+      dotColor: dotColor,
+    };
+  }
+
+  return {dates: datesForCalendar};
 };
 
 export const getMarkedToday = ({colorToday}: IColors) => {